ISSUE #2: Implement Flush for AAC-Encoder
This is not yet implemented correctly. According to the documentation, setting the numInSampes to -1 should be sufficient but that's not entirely
the answer.
When queining up multiple WAV files to be converted, every file after the first one has a blip at the beginning. This doesn't happen for the first
file. This bug existed for the MPEG Encoder as well but implementing the MpegEncoder::DoFlush() fixed that bug.
The audio blip at the beginning sounds like some remnant of an earlier buffer.
Further investigation is required.