Template the DailyWorker on a WaitFunctor
WaitFunctor() returns bool. client-code uses WaitFunctor to tell DailyWorker if it can proceed with the scheduledTask.
The client-code will setup WaitFunctor such that it returns false when the other queues (EncoderQueue, DownloadQueue, PlaybackQueue) are busy.
This could have been done easily with a std::function<bool()> but the functor will be executed within a while loop, hence it's templated.